pyproject.toml
[project]
name = "ai-kanban"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "notion-client>=2.4.0",
    "pika>=1.3.2",
    "python-dotenv>=1.1.1",
    "anthropic>=0.40.0",
    "aio-pika>=9.4.0",
    "openai>=1.93.1",
    "requests>=2.32.4",
    "langgraph>=0.0.60",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.8.0",
    "mypy>=1.16.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88
extend-exclude = [
    ".git",
    ".pytest_cache",
    ".venv",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "E203",  # whitespace before ':' (conflicts with black)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "F401",  # unused imports in tests are common
]
